requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
cryptography==44.0.2
Django==5.1.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
pycparser==2.22
pycryptodome==3.11.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
requests==2.26.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==1.26.20
